tests: Move common_sequential to integration_common_sequential.rs
Extract mod common_sequential from integration.rs into a
standalone test binary integration_common_sequential.rs.
This follows the pattern established by integration_cvm.rs.
